결제 뭔가 되고있다...!! 하하하하하하하하하

+++ matdori/src/main/java/com/kh/matdori/vo/MenuInfoVO.java
@@ -11,6 +11,6 @@ public class MenuInfoVO {
 	private int resNo;
 	private String menuName;
 	private int menuQty;
-	private float menuPrice;
+	private int menuPrice;
 
 }
